pyproject.toml
[project]
name = "gym_continual_rl"
version = "0.1.0"
dependencies = [
  "gymnasium==0.29.1",
  "minigrid==2.3.1",
  "numpy==1.26.4",
  "pygame==2.5.2"
]

[project.optional-dependencies]
test = [
  "black==24.4.2",
  "flake8==7.0.0",
  "isort==5.13.2",
  "pillow==10.3.0",
  "pytest==8.2.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.isort]
profile = "black"
